msm: kgsl: Correctly handle dcvs hfi failures
For dcvs hfi failure that happen alongside an active
gpu, do an inline snapshot. Because we need to reset
GMU, we need to reset GPU as well. So set the dispatcher
fault and trigger dispatcher for recovery.

Also, treat the dcvs failure during GMU boot just
like other GMU boot failures so that GMU is reset.

Change-Id: I76dd08e129fa74a127e1795fd9de8e03612fe666
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>